good steps for updates and deletes.  I've got both the pred-succ cache and the oaiid-recordid cache working.  I've verified that it works for both the DefaultRepository and the TestRepository.
git-svn-id: https://xcmetadataservicestoolkit.googlecode.com/svn/branches/bens_perma_branch@934 72d1a826-0e72-11de-b5bf-434079a3519f

+++ mst-common/src/java/xc/mst/repo/sql/create_repo.sql
@@ -33,9 +33,9 @@ create table repo_name.record_sets (
 ) ENGINE=MyISAM DEFAULT CHARSET=utf8;
 
 create table repo_name.record_predecessors (
-	id                    int        NOT NULL    AUTO_INCREMENT,
-	record_id             int,
-	pred_record_id        int,
+	id                    BIGINT        NOT NULL    AUTO_INCREMENT,
+	record_id             BIGINT,
+	pred_record_id        BIGINT,
 	PRIMARY KEY (id),
 	KEY idx_REPO_NAME_records_set_record_id(record_id),
 	KEY idx_REPO_NAME_records_set_pred_record_id(pred_record_id)
